==> 00_spc-vip-graph.R <==
################################################################################
## Create a stacked graph of raw spectra, preprocessed spectra and 
## variable importance in the projection (VIP) scores for 
## the total C PLSR model, total N PLSR model and clay PLSR model
################################################################################
